backend/pyproject.toml
[project]
name = "backend-py"
version = "0.1.0"
description = "AudioDiary Backend API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.121.0",
    "uvicorn[standard]>=0.38.0",
    "python-dotenv>=1.2.0",
    "google-generativeai>=0.8.0",
]


[project.optional-dependencies]
dev = [
    "ruff>=0.14.0",
    "mypy>=1.18.0",
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "N",
    "W",
    "UP",
    "B",
    "C4",
    "PIE",
    "SIM",
    "T20",
    "PT",
    "Q",
    "RUF",
]
ignore = [
    "E501",
    "B008",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "google.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
